[Silver V] Title: 귀찮음, Time: 640 ms, Memory: 56084 KB, Score: 30 point -BaekjoonHub

+++ 백준/Silver/16208. 귀찮음/귀찮음.java
@@ -18,7 +18,7 @@ public class Main {
             int one = heap.poll();
             int two = heap.poll();
 
-            total += one * two;
+            total += (long)one * two;
             heap.add(one+two);
         }
